requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.32.1

# Database
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0
aiosqlite==0.20.0

# Data validation
pydantic==2.9.2
pydantic-settings==2.6.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7
python-multipart==0.0.6

# Environment & Configuration
python-dotenv==1.0.0
click==8.1.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# File handling & Security
python-magic==0.4.27
Pillow==10.4.0

# HTTP client
aiohttp==3.11.10

# Caching & Performance
redis==5.2.0
aiocache==0.12.3
asyncio-throttle==1.0.2
cachetools==5.5.0

# Monitoring & Profiling
psutil==6.1.0
prometheus-client==0.21.0

# Utilities
tabulate==0.9.0 